redefine outpoint to be an opaque 36 byte array
While in actual bitcoin transactions the txid and index matter, here we treat the whole outpoint as a unique identifier.
It should be simpler to treat them as arrays since we shouldn't need to look at the separate elements.
No data is lost so if we want to show the txid:idx we can, for example when we need to ask bitcoind about it.