deduplicate outpoint struct, move validator to database package
outpoint was defined in both message and database.  Outpoint feels more like a message thing
than a database specific thing, and making it a message only looked like it would cause an
import cycle.  Moving validator to database fixes that, and validator "feels" more like part
of the database and not part of a message library.  This seems in keeping with btcd or lnd
"wire" packages at least.